python/requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
attrs==21.2.0
black==21.4b2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
cloudpickle==2.0.0
cycler==0.11.0
Cython==0.29.24
-e git+https://github.com/facebookresearch/detectron2.git@8745a31749f9e90f108d330622edea60e15d0ddd#egg=detectron2
fonttools==4.28.2
future==0.18.2
fvcore==0.1.5.post20211023
gevent==21.8.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.42.0
hydra-core==1.1.1
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
iopath==0.1.9
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.0
mypy-extensions==0.4.3
numpy==1.21.3
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.4.60
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
pluggy==1.0.0
portalocker==2.3.2
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pydot==1.4.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
setuptools-scm==6.3.2
six==1.16.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
yacs==0.1.8
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
